+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/sound/SoundFormat.java
@@ -224,7 +224,78 @@ public class SoundFormat {
         return decodedData;
     }
     
-    public boolean createWav(SOUNDINFO soundInfo, List<ByteArrayRange> dataRanges, OutputStream os, int skipSamples) throws IOException {
+    private byte[] resample(byte[] decodedData) throws IOException {
+        if (samplingRate == 44100) {
+            return decodedData;
+        }
+        boolean resamplingFromStereo = true;
+        
+        ByteArrayOutputStream baosResampled = new ByteArrayOutputStream();
+        for (int i = 0; i < decodedData.length; i += (resamplingFromStereo ? 4 : 2)) {
+            if (i + 1 >= decodedData.length) {
+                break;
+            }
+            int left = ((decodedData[i] & 0xff) + ((decodedData[i + 1] & 0xff) << 8)) << 16 >> 16;
+            int right = left;
+            if (resamplingFromStereo) 
+            {
+                if (i + 3 >= decodedData.length) {
+                    break;
+                }
+                right = ((decodedData[i + 2] & 0xff) + ((decodedData[i + 3] & 0xff) << 8)) << 16 >> 16;
+            }
+
+            int nextLeft = left;
+            int nextRight = right;
+            int nextI = i + (resamplingFromStereo ? 4 : 2);
+            if (nextI < decodedData.length) {
+                nextLeft = ((decodedData[nextI] & 0xff) + ((decodedData[nextI + 1] & 0xff) << 8)) << 16 >> 16;
+                nextRight = nextLeft;
+                if (resamplingFromStereo) {
+                    if (nextI + 3 >= decodedData.length) {
+                        //ignore
+                    } else {
+                        nextRight = ((decodedData[nextI + 2] & 0xff) + ((decodedData[nextI + 3] & 0xff) << 8)) << 16 >> 16;
+                    }
+                }
+            }                
+
+            writeLE(baosResampled, left, 2);
+            writeLE(baosResampled, right, 2);
+            if (samplingRate == 5512) {
+                writeLE(baosResampled, left + (nextLeft - left) / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 2 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 2 / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 3 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 3 / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 4 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 4 / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 5 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 5 / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 6 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 6 / 8, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 7 / 8, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 7 / 8, 2);
+            }
+            if (samplingRate == 11025) {
+                writeLE(baosResampled, left + (nextLeft - left) / 4, 2);
+                writeLE(baosResampled, right + (nextRight - right) / 4, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 2 / 4, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 2 / 4, 2);
+                writeLE(baosResampled, left + (nextLeft - left) * 3 / 4, 2);
+                writeLE(baosResampled, right + (nextRight - right) * 3 / 4, 2);
+            }
+            if (samplingRate == 22050)
+            {
+                writeLE(baosResampled, (left + nextLeft) / 2, 2);
+                writeLE(baosResampled, (right + nextRight) / 2, 2);
+            }            
+        }
+        return baosResampled.toByteArray();
+    }
+    
+    public boolean createWav(SOUNDINFO soundInfo, List<ByteArrayRange> dataRanges, OutputStream os, int skipSamples, boolean resample) throws IOException {
         
         byte[] decodedData = decode(soundInfo, dataRanges, skipSamples);
         boolean convertedStereo = stereo;
@@ -252,8 +323,8 @@ public class SoundFormat {
                         break;
                     }
                     right = ((decodedData[i + 2] & 0xff) + ((decodedData[i + 3] & 0xff) << 8)) << 16 >> 16;
-                }
-
+                }                                
+                
                 if (soundInfo.hasEnvelope) {
                     for (int e = 0; e < soundInfo.envelopeRecords.length - 1; e++) {
                         int envPosBytes = inPointBytes + (int) (soundInfo.envelopeRecords[e].pos44 * samplingRate / 44100.0 * 2 * (stereo ? 2 : 1));
@@ -274,13 +345,16 @@ public class SoundFormat {
                 }
 
                 writeLE(baosFiltered, left, 2);
-                writeLE(baosFiltered, right, 2);
+                writeLE(baosFiltered, right, 2);                    
             }
             convertedStereo = true;
         }
+        
+        
+        byte[] resampled = resample ? resample(baosFiltered.toByteArray()) : baosFiltered.toByteArray();
 
         try {
-            createWavFromPcmData(os, samplingRate, true, convertedStereo, baosFiltered.toByteArray());
+            createWavFromPcmData(os, resample ? 44100 : samplingRate, true, convertedStereo, resampled);
             return true;
         } catch (IOException ex) {
             return false;